Added libzvbi and libiconv. Edited for clarity.
Added libzvbi and its pre-requisite libiconv in order to handle DVB streams with embedded teletext subtitles.

Unfortunately, this doesn't mean that the subtitles will work, just that hacktv won't crash with an "Unsupported subtitle codec" error when trying to open the stream. This is consistent with the behaviour on non-Windows systems.